# api/openapi.yaml
openapi: 3.1.0
info:
  title: tyrs-hand Management API
  version: 0.1.0
  license:
    name: MIT
    identifier: MIT
servers:
  - url: /api/v1
...
paths:
  /webhooks/github:
    post:
      operationId: receiveGitHubWebhook
      tags: [GitHub]
      servers: [{ url: / }]
      security: []
      parameters:
        - { name: X-Hub-Signature-256, in: header, required: true, schema: { type: string } }
        - { name: X-GitHub-Delivery, in: header, required: true, schema: { type: string } }
        - { name: X-GitHub-Event, in: header, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        "202":
          description: 已持久化
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        default: { $ref: "#/components/responses/Problem" }
  /internal/v1/tools/call:
    post:
      operationId: callInternalTool
      tags: [Resources]
      servers: [{ url: / }]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        "200":
          description: Tool Call 结果
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        default: { $ref: "#/components/responses/Problem" }
...
  /setup/admin:
    post:
      operationId: setupAdministrator
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SetupRequest" }
      responses:
        "201":
          description: 管理员与恢复资料
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SetupResult" }
        default: { $ref: "#/components/responses/Problem" }
  /auth/login:
    post:
      operationId: login
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: 登录会话
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResult" }
        default: { $ref: "#/components/responses/Problem" }
  /auth/me:
    get:
      operationId: getCurrentAdministrator
      tags: [Auth]
      responses:
        "200":
          description: 当前管理员
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Administrator" }
        default: { $ref: "#/components/responses/Problem" }
  /auth/logout:
    post:
      operationId: logout
      tags: [Auth]
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }]
      responses:
        "204": { description: 已退出 }
        default: { $ref: "#/components/responses/Problem" }
  /github/app:
    get:
      operationId: getGitHubApp
      tags: [GitHub]
      responses:
        "200":
          description: GitHub App 状态
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GitHubAppStatus" }
    put:
      operationId: putGitHubApp
      tags: [GitHub]
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GitHubAppInput" }
      responses:
        "204": { description: 已保存 }
        default: { $ref: "#/components/responses/Problem" }
...
  /github/app/manifest/callback:
    get:
      operationId: githubAppManifestCallback
      tags: [GitHub]
      parameters:
        - name: code
          in: query
          required: true
          schema: { type: string }
      responses:
        "302": { description: 返回 GitHub 设置页 }
        default: { $ref: "#/components/responses/Problem" }
  /settings/global-agents:
    get:
      operationId: getGlobalAgents
      tags: [Settings]
      responses:
        "200":
          description: 作为任务 developer instructions 注入的全局 Agent 指令
          content: { application/json: { schema: { $ref: "#/components/schemas/GlobalAgents" } } }
        default: { $ref: "#/components/responses/Problem" }
    put:
      operationId: putGlobalAgents
      tags: [Settings]
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/GlobalAgents" } } }
      responses:
        "204": { description: 已保存 }
        default: { $ref: "#/components/responses/Problem" }
  /settings/codex:
    get:
      operationId: listCodexSettings
      tags: [Settings]
      responses:
        "200":
          description: 仓库与 Discord Forum 的 Codex 设置
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CodexSettingsList" }
        default: { $ref: "#/components/responses/Problem" }
  /settings/codex/repositories/{id}:
    put:
      operationId: putRepositoryCodexSettings
      tags: [Settings]
      parameters:
        - { $ref: "#/components/parameters/CSRFToken" }
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CodexPreferences" } } }
      responses:
        "204": { description: 已保存 }
        default: { $ref: "#/components/responses/Problem" }
  /settings/codex/forums/{id}:
    put:
      operationId: putForumCodexSettings
      tags: [Settings]
      parameters:
        - { $ref: "#/components/parameters/CSRFToken" }
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CodexPreferences" } } }
      responses:
        "204": { description: 已保存 }
        default: { $ref: "#/components/responses/Problem" }
  /settings/discord:
    get:
      operationId: getDiscordSettings
      tags: [Discord]
      responses:
        "200":
          description: Discord 设置
          content: { application/json: { schema: { $ref: "#/components/schemas/DiscordSettings" } } }
        default: { $ref: "#/components/responses/Problem" }
    put:
      operationId: putDiscordSettings
      tags: [Discord]
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/DiscordSettingsInput" } } }
      responses:
        "204": { description: 已保存 }
        default: { $ref: "#/components/responses/Problem" }
  /discord/status:
    get:
      operationId: getDiscordStatus
      tags: [Discord]
      responses:
        "200":
          description: Discord 运行状态
          content: { application/json: { schema: { $ref: "#/components/schemas/DiscordStatus" } } }
        default: { $ref: "#/components/responses/Problem" }
  /discord/initializations/preflight:
    post:
      operationId: preflightDiscordInitialization
      tags: [Discord]
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/DiscordInitializationInput" } } }
      responses:
        "200":
          description: 零修改预检结果
          content: { application/json: { schema: { $ref: "#/components/schemas/DiscordPreflight" } } }
        default: { $ref: "#/components/responses/Problem" }
  /discord/initializations:
    post:
      operationId: createDiscordInitialization
      tags: [Discord]
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/DiscordInitializationInput" } } }
      responses:
        "202":
          description: 初始化已排队
          content: { application/json: { schema: { $ref: "#/components/schemas/IDResource" } } }
        default: { $ref: "#/components/responses/Problem" }
  /discord/initializations/{id}:
    get:
      operationId: getDiscordInitialization
      tags: [Discord]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: 初始化状态
          content: { application/json: { schema: { $ref: "#/components/schemas/DiscordInitialization" } } }
        default: { $ref: "#/components/responses/Problem" }
  /discord/members:
    get:
      operationId: listDiscordMembers
      tags: [Discord]
      responses:
        "200":
          description: Discord 成员
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/DiscordMember" } }
        default: { $ref: "#/components/responses/Problem" }
  /development-environments:
    get:
      operationId: listDevelopmentEnvironments
      tags: [Development Environments]
      responses:
        "200":
          description: 长期开发环境
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DevelopmentEnvironmentList" }
        default: { $ref: "#/components/responses/Problem" }
    post:
      operationId: createDevelopmentEnvironment
      tags: [Development Environments]
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ownerDiscordUserId]
              properties:
                ownerDiscordUserId: { type: string, minLength: 1 }
      responses:
        "201":
          description: 逻辑环境已创建并绑定 Worker
          content: { application/json: { schema: { $ref: "#/components/schemas/IDResource" } } }
        default: { $ref: "#/components/responses/Problem" }
  /development-environments/{id}/ssh:
    put:
      operationId: putDevelopmentEnvironmentSSH
      tags: [Development Environments]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { $ref: "#/components/parameters/CSRFToken" }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/DevelopmentEnvironmentSSHInput" } } }
      responses:
        "202": { description: SSH 配置已排队 }
        default: { $ref: "#/components/responses/Problem" }
    delete:
      operationId: deleteDevelopmentEnvironmentSSH
      tags: [Development Environments]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { $ref: "#/components/parameters/CSRFToken" }
      responses:
        "202": { description: SSH 停用已排队 }
        default: { $ref: "#/components/responses/Problem" }
  /development-projects/{id}/forums:
    post:
      operationId: createDevelopmentProjectForum
      tags: [Development Environments]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { $ref: "#/components/parameters/CSRFToken" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DevelopmentProjectForumInput" }
      responses:
        "202": { description: Forum 创建已排队或历史 Forum 已恢复 }
        default: { $ref: "#/components/responses/Problem" }
  /development-forums/{id}/disable:
    post:
      operationId: disableDevelopmentForum
      tags: [Development Environments]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { $ref: "#/components/parameters/CSRFToken" }
      responses:
        "202": { description: Forum 配对已停用并切换为只读 }
        default: { $ref: "#/components/responses/Problem" }
  /development-forums/{id}/enable:
    post:
      operationId: enableDevelopmentForum
      tags: [Development Environments]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { $ref: "#/components/parameters/CSRFToken" }
      responses:
        "202": { description: Forum 配对已恢复 }
        default: { $ref: "#/components/responses/Problem" }
  /development-projects/{id}/forums/{forumId}/collaborators/{memberId}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      - { name: forumId, in: path, required: true, schema: { type: string, format: uuid } }
      - { name: memberId, in: path, required: true, schema: { type: string } }
    put:
      operationId: putDevelopmentProjectForumCollaborator
      tags: [Development Environments]
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [accessLevel]
              properties: { accessLevel: { type: string, enum: [readonly, operator] } }
      responses:
        "204": { description: 已更新 }
        default: { $ref: "#/components/responses/Problem" }
    delete:
      operationId: deleteDevelopmentProjectForumCollaborator
      tags: [Development Environments]
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }]
      responses:
        "204": { description: 已删除 }
        default: { $ref: "#/components/responses/Problem" }
  /discord/forums/{forumId}/access/{memberId}:
    parameters:
      - { name: forumId, in: path, required: true, schema: { type: string, format: uuid } }
      - { name: memberId, in: path, required: true, schema: { type: string } }
    put:
      operationId: putDiscordForumAccess
      tags: [Discord]
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [accessLevel]
              properties: { accessLevel: { type: string, enum: [readonly, operator] } }
      responses:
        "204": { description: 已更新 }
        default: { $ref: "#/components/responses/Problem" }
    delete:
      operationId: deleteDiscordForumAccess
      tags: [Discord]
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }]
      responses:
        "204": { description: 已删除 }
        default: { $ref: "#/components/responses/Problem" }
  /discord/github/bind:
    post:
      operationId: startDiscordGitHubBind
      tags: [Discord]
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [guildId, discordUserId]
              properties: { guildId: { type: string }, discordUserId: { type: string } }
      responses:
        "200":
          description: 一次性绑定链接
          content:
            application/json:
              schema:
                type: object
                required: [url]
                properties: { url: { type: string, format: uri } }
        default: { $ref: "#/components/responses/Problem" }
  /discord/github/bind/callback:
    get:
      operationId: completeDiscordGitHubBind
      tags: [Discord]
      security: []
      parameters:
        - { name: state, in: query, required: true, schema: { type: string } }
        - { name: code, in: query, required: true, schema: { type: string } }
      responses:
        "200": { description: 绑定完成 }
        default: { $ref: "#/components/responses/Problem" }
  /discord/github/unbind:
    post:
      operationId: unbindDiscordGitHub
      tags: [Discord]
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [guildId, discordUserId, confirmed]
              properties:
                guildId: { type: string }
                discordUserId: { type: string }
                confirmed: { type: boolean }
      responses:
        "204": { description: 已解绑 }
        default: { $ref: "#/components/responses/Problem" }
...
  /repositories:
    get: { $ref: "#/components/pathItems/ListRepositories" }
    post:
      operationId: createRepository
      tags: [Resources]
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RepositoryInput" }
      responses:
        "201":
          description: 已创建
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IDResource" }
        default: { $ref: "#/components/responses/Problem" }
...
  /trigger-rules:
    get: { $ref: "#/components/pathItems/ListTriggerRules" }
    post:
      operationId: createTriggerRule
      tags: [Resources]
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TriggerRuleInput" }
      responses:
        "201":
          description: 已创建
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IDResource" }
        default: { $ref: "#/components/responses/Problem" }
...
  /workers:
    get:
      operationId: listHostWorkers
      tags: [Workers]
      responses:
        "200":
          description: 宿主 Worker 列表
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties: { items: { type: array, items: { $ref: "#/components/schemas/ExecutionNode" } } }
    post:
      operationId: createHostWorker
      tags: [Workers]
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/ExecutionNodeInput" } } }
      responses:
        "201":
          description: Worker 与一次性注册 Token
          content: { application/json: { schema: { $ref: "#/components/schemas/ExecutionNodeCreated" } } }
  /workers/{id}/enrollments:
    post:
      operationId: createHostWorkerEnrollment
      tags: [Workers]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { $ref: "#/components/parameters/CSRFToken" }
      responses:
        "201": { description: 一次性注册 Token, content: { application/json: { schema: { $ref: "#/components/schemas/EnrollmentToken" } } } }
  /workers/{id}:
    delete:
      operationId: deleteHostWorker
      tags: [Workers]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { $ref: "#/components/parameters/CSRFToken" }
      responses: { "204": { description: 已删除 }, default: { $ref: "#/components/responses/Problem" } }
  /workers/{id}/enabled:
    put:
      operationId: setHostWorkerEnabled
      tags: [Workers]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { $ref: "#/components/parameters/CSRFToken" }
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [enabled], properties: { enabled: { type: boolean } } } } } }
      responses: { "204": { description: 已更新 }, default: { $ref: "#/components/responses/Problem" } }
...
  /system/status:
    get:
      operationId: getSystemStatus
      tags: [Resources]
      responses:
        "200":
          description: 系统状态
          content:
            application/json:
              schema: { type: object, additionalProperties: { type: integer, format: int64 } }
        default: { $ref: "#/components/responses/Problem" }
...
  /client/device:
    delete:
      operationId: clientDeleteDevice
      tags: [Client Protocol]
      security: [{ clientBearer: [] }]
      responses: { "204": { description: 当前设备已自撤销 }, default: { $ref: "#/components/responses/Problem" } }
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: tyrs_hand_session
    nodeBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-node-token
    clientBearer:
      type: http
      scheme: bearer
      bearerFormat: tyrs-hand-device-token
  parameters:
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string }
...
  responses:
    Problem:
      description: RFC 9457 Problem Details
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetails" }
...
  schemas:
...
    SetupRequest:
      type: object
      required: [setupToken, username, password]
      properties:
        setupToken: { type: string, minLength: 16 }
        username: { type: string, minLength: 3, maxLength: 64 }
        password: { type: string, minLength: 12, maxLength: 256 }
    SetupResult:
      type: object
      required: [totpSecret, provisioningUri, recoveryCodes]
      properties:
        totpSecret: { type: string }
        provisioningUri: { type: string, format: uri }
        recoveryCodes: { type: array, items: { type: string } }
    LoginRequest:
      type: object
      required: [username, password, totp]
      properties:
        username: { type: string }
        password: { type: string }
        totp: { type: string, pattern: "^[0-9]{6}$" }
...
security:
  - sessionCookie: []