# internal/httpapi/router.go
package httpapi

import (
	"context"
	"database/sql"
	"io/fs"
	"net/http"
	"strconv"
	"strings"
	"time"

	"github.com/gin-gonic/gin"
	"github.com/google/uuid"
	"github.com/prometheus/client_golang/prometheus/promhttp"
	"github.com/redis/go-redis/v9"
	"github.com/slovx2/tyrs-hand/internal/auth"
	"github.com/slovx2/tyrs-hand/internal/config"
	"github.com/slovx2/tyrs-hand/internal/discordintegration"
	"github.com/slovx2/tyrs-hand/internal/executionnode"
	ghadapter "github.com/slovx2/tyrs-hand/internal/github"
	"github.com/slovx2/tyrs-hand/internal/githubtools"
	"github.com/slovx2/tyrs-hand/internal/secrets"
	platformsettings "github.com/slovx2/tyrs-hand/internal/settings"
	"github.com/slovx2/tyrs-hand/internal/sshconfig"
	"github.com/slovx2/tyrs-hand/internal/web"
	"go.uber.org/zap"
)

const sessionCookie = "tyrs_hand_session"

type Server struct {
	cfg                config.Config
	db                 *sql.DB
	redis              *redis.Client
	auth               *auth.Service
	github             *ghadapter.Manager
	catalog            *githubtools.Catalog
	settings           *platformsettings.Service
	discord            *discordintegration.Manager
	desktopImageRemote func(context.Context) (desktopImageDiscord, error)
	bindings           *discordintegration.BindingService
	nodes              *executionnode.Service
	ssh                *sshconfig.Service
	secrets            *secrets.Store
	logger             *zap.Logger
	assets             fs.FS
	clientUpdateHub    *clientUpdateHub
}

func NewServer(cfg config.Config, db *sql.DB, redisClient *redis.Client, authService *auth.Service, githubManager *ghadapter.Manager, catalog *githubtools.Catalog, settingsService *platformsettings.Service, discordManager *discordintegration.Manager, bindingService *discordintegration.BindingService, secretStore *secrets.Store, logger *zap.Logger) (*Server, error) {
	assets, err := web.Assets()
	if err != nil {
		return nil, err
	}
	return &Server{cfg: cfg, db: db, redis: redisClient, auth: authService, github: githubManager,
		catalog: catalog, settings: settingsService, discord: discordManager, bindings: bindingService,
		nodes: executionnode.NewService(db), ssh: sshconfig.NewService(db, secretStore),
		secrets: secretStore, logger: logger, assets: assets,
		clientUpdateHub: newClientUpdateHub()}, nil
}

func (s *Server) baseRouter() *gin.Engine {
	if s.cfg.Environment == "production" {
		gin.SetMode(gin.ReleaseMode)
	}
	router := gin.New()
	_ = router.SetTrustedProxies(nil)
	router.Use(s.requestID(), s.securityHeaders(), s.rateLimit(), gin.Recovery(), metricsMiddleware(), s.accessLog())
	router.GET("/healthz", func(c *gin.Context) { c.JSON(http.StatusOK, gin.H{"status": "ok"}) })
	router.GET("/readyz", s.ready)
	return router
}

func (s *Server) Router() http.Handler {
	return s.adminRouter(true)
}

func (s *Server) AdminRouter() http.Handler {
	return s.adminRouter(false)
}

func (s *Server) adminRouter(includeWebhook bool) http.Handler {
	router := s.baseRouter()
	router.GET("/metrics", func(c *gin.Context) {
		s.refreshOperationalMetrics(c.Request.Context())
		gin.WrapH(promhttp.Handler())(c)
	})
	router.POST("/internal/v1/tools/call", s.internalToolCall)
	router.POST("/internal/v1/tools/failure", s.internalToolFailure)
	router.POST("/internal/v1/git/credential", s.internalGitCredential)
	s.registerWorkerRoutes(router)
	if includeWebhook {
		router.POST("/webhooks/github", s.githubWebhook)
	}

	api := router.Group("/api/v1")
	api.GET("/setup/status", s.setupStatus)
	api.POST("/setup/admin", s.setupAdmin)
	api.POST("/auth/login", s.login)
	api.POST("/client/auth/login", s.clientLogin)
	api.POST("/client/device-pairings/:id/claim", s.claimClientDevicePairing)
	api.GET("/client/device-pairings/:id/status", s.clientDevicePairingStatus)
	api.GET("/github/app/manifest/callback", s.githubManifestCallback)
	api.GET("/discord/github/bind/callback", s.discordGitHubBindCallback)
	client := api.Group("/client")
	client.Use(s.requireClientBearer())
	client.GET("/bootstrap", s.clientBootstrap)
	client.GET("/sessions", s.clientListSessions)
	client.POST("/sessions", s.clientCreateSession)
	client.GET("/sessions/:id", s.clientGetSession)
	client.PATCH("/sessions/:id", s.clientPatchSession)
	client.POST("/sessions/:id/stop", s.clientStopSession)
	client.POST("/sessions/:id/archive", s.clientArchiveSession)
	client.POST("/sessions/:id/restore", s.clientRestoreSession)
	client.POST("/sessions/:id/plans/:runId/execute", s.clientExecutePlan)
	client.GET("/sessions/:id/messages", s.clientListMessages)
	client.POST("/sessions/:id/messages", s.clientCreateMessage)
	client.POST("/uploads", s.clientUpload)
	client.GET("/attachments/:id/content", s.clientDownloadAttachment)
	client.POST("/interactive/:id/answer", s.clientAnswerInteractive)
	client.PUT("/device/push-token", s.clientPutPushToken)
	client.DELETE("/device/push-token", s.clientDeletePushToken)
	client.DELETE("/device", s.clientDeleteDevice)
	client.GET("/sync", s.clientSync)
	client.GET("/updates", s.clientUpdates)

	authenticated := api.Group("")
	authenticated.Use(s.requireSession())
	authenticated.GET("/auth/me", s.me)
	authenticated.POST("/auth/logout", s.requireCSRF(), s.logout)
	authenticated.GET("/github/app", s.getGitHubApp)
	authenticated.PUT("/github/app", s.requireCSRF(), s.putGitHubApp)
	authenticated.GET("/github/app/manifest", s.githubManifest)
	authenticated.GET("/repositories", s.listRepositories)
	authenticated.GET("/installations", s.listInstallations)
	authenticated.POST("/repositories", s.requireCSRF(), s.createRepository)
	authenticated.GET("/agent-profiles", s.listAgentProfiles)
	authenticated.POST("/agent-profiles", s.requireCSRF(), s.createAgentProfile)
	authenticated.GET("/trigger-rules", s.listTriggerRules)
	authenticated.POST("/trigger-rules", s.requireCSRF(), s.createTriggerRule)
	authenticated.GET("/work-items", s.listWorkItems)
	authenticated.GET("/jobs", s.listJobs)
	authenticated.GET("/workers", s.listExecutionNodes)
	authenticated.GET("/client-devices", s.listClientDevices)
	authenticated.DELETE("/client-devices/:id", s.requireCSRF(), s.deleteClientDevice)
	authenticated.POST("/client-device-pairings", s.requireCSRF(), s.createClientDevicePairing)
	authenticated.GET("/client-device-pairings/:id", s.getClientDevicePairing)
	authenticated.POST("/client-device-pairings/:id/approve", s.requireCSRF(), s.approveClientDevicePairing)
	authenticated.POST("/client-device-pairings/:id/reject", s.requireCSRF(), s.rejectClientDevicePairing)
	authenticated.POST("/workers", s.requireCSRF(), s.createExecutionNode)
	authenticated.POST("/workers/:id/enrollments", s.requireCSRF(), s.createExecutionNodeEnrollment)
	authenticated.PUT("/workers/:id/enabled", s.requireCSRF(), s.setExecutionNodeEnabled)
	authenticated.DELETE("/workers/:id", s.requireCSRF(), s.deleteExecutionNode)
	authenticated.GET("/ssh/credentials", s.listSSHCredentials)
	authenticated.POST("/ssh/credentials", s.requireCSRF(), s.createSSHCredential)
	authenticated.PUT("/ssh/credentials/:id", s.requireCSRF(), s.updateSSHCredential)
	authenticated.DELETE("/ssh/credentials/:id", s.requireCSRF(), s.deleteSSHCredential)
	authenticated.GET("/ssh/hosts", s.listSSHHosts)
	authenticated.POST("/ssh/hosts", s.requireCSRF(), s.createSSHHost)
	authenticated.POST("/ssh/hosts/import", s.requireCSRF(), s.importSSHHosts)
	authenticated.PUT("/ssh/hosts/:id", s.requireCSRF(), s.updateSSHHost)
	authenticated.DELETE("/ssh/hosts/:id", s.requireCSRF(), s.deleteSSHHost)
	authenticated.GET("/settings/execution", s.getExecutionSettings)
	authenticated.PUT("/settings/execution", s.requireCSRF(), s.putExecutionSettings)
	authenticated.GET("/threads", s.listThreads)
	authenticated.POST("/controls/:id/reconcile", s.requireCSRF(), s.reconcileControl)
	authenticated.POST("/controls/:id/reset", s.requireCSRF(), s.resetControl)
	authenticated.GET("/worktrees", s.listWorktrees)
	authenticated.GET("/repo-caches", s.listRepoCaches)
	authenticated.GET("/audit-logs", s.listAuditLogs)
	authenticated.GET("/settings/global-agents", s.getGlobalAgents)
	authenticated.PUT("/settings/global-agents", s.requireCSRF(), s.putGlobalAgents)
	authenticated.GET("/settings/codex", s.listCodexSettings)
	authenticated.PUT("/settings/codex/repositories/:id", s.requireCSRF(), s.putRepositoryCodexSettings)
	authenticated.PUT("/settings/codex/forums/:id", s.requireCSRF(), s.putForumCodexSettings)
	authenticated.GET("/settings/discord", s.getDiscordSettings)
	authenticated.PUT("/settings/discord", s.requireCSRF(), s.putDiscordSettings)
	authenticated.GET("/discord/status", s.discordStatus)
	authenticated.POST("/discord/initializations/preflight", s.requireCSRF(), s.discordInitializationPreflight)
	authenticated.POST("/discord/initializations", s.requireCSRF(), s.createDiscordInitialization)
	authenticated.GET("/discord/initializations/:id", s.getDiscordInitialization)
	authenticated.GET("/discord/members", s.listDiscordMembers)
	authenticated.GET("/development-environments", s.listDevelopmentEnvironments)
	authenticated.POST("/development-environments", s.requireCSRF(), s.createDevelopmentEnvironment)
	authenticated.PUT("/development-environments/:id/ssh", s.requireCSRF(), s.putDevelopmentEnvironmentSSH)
	authenticated.DELETE("/development-environments/:id/ssh", s.requireCSRF(), s.deleteDevelopmentEnvironmentSSH)
	authenticated.POST("/development-projects/:id/forums", s.requireCSRF(), s.createDevelopmentProjectForum)
	authenticated.POST("/development-forums/:id/disable", s.requireCSRF(), s.disableDevelopmentForum)
	authenticated.POST("/development-forums/:id/enable", s.requireCSRF(), s.enableDevelopmentForum)
	authenticated.PUT("/development-projects/:id/forums/:forumId/collaborators/:memberId",
		s.requireCSRF(), s.putDevelopmentProjectForumCollaborator)
	authenticated.DELETE("/development-projects/:id/forums/:forumId/collaborators/:memberId",
		s.requireCSRF(), s.deleteDevelopmentProjectForumCollaborator)
	authenticated.PUT("/discord/forums/:forumId/access/:memberId", s.requireCSRF(), s.putDiscordForumAccess)
	authenticated.DELETE("/discord/forums/:forumId/access/:memberId", s.requireCSRF(), s.deleteDiscordForumAccess)
	authenticated.POST("/discord/github/bind", s.requireCSRF(), s.startDiscordGitHubBind)
	authenticated.POST("/discord/github/unbind", s.requireCSRF(), s.unbindDiscordGitHub)
	authenticated.GET("/system/status", s.systemStatus)
	authenticated.GET("/events/stream", s.eventsStream)

	router.NoRoute(s.serveSPA)
	return router
}

func (s *Server) WebhookRouter() http.Handler {
	router := s.baseRouter()
	router.POST("/webhooks/github", s.githubWebhook)
	return router
}

func (s *Server) rateLimit() gin.HandlerFunc {
	return func(c *gin.Context) {
		bucketName, limit := rateLimitPolicy(c.Request.URL.Path)
		bucket := time.Now().UTC().Unix() / 60
		key := "tyrs-hand:rate:" + bucketName + ":" + c.ClientIP() + ":" + strconv.FormatInt(bucket, 10)
		count, err := s.redis.Incr(c.Request.Context(), key).Result()
		if err == nil {
			if count == 1 {
				_ = s.redis.Expire(c.Request.Context(), key, 2*time.Minute).Err()
			}
			c.Header("X-RateLimit-Limit", strconv.FormatInt(limit, 10))
			c.Header("X-RateLimit-Remaining", strconv.FormatInt(max(0, limit-count), 10))
			if count > limit {
				problem(c, http.StatusTooManyRequests, "请求过于频繁", nil)
				return
			}
		}
		c.Next()
	}
}

func rateLimitPolicy(path string) (string, int64) {
	if strings.HasPrefix(path, "/worker/v2/") {
		return "worker-api", 10000
	}
	switch path {
	case "/api/v1/auth/login":
		return "auth-login", 10
	case "/api/v1/setup/admin":
		return "setup-admin", 10
	case "/webhooks/github":
		return "github-webhook", 300
	default:
		return "default", 600
	}
}

func (s *Server) ready(c *gin.Context) {
	ctx, cancel := context.WithTimeout(c.Request.Context(), 2*time.Second)
	defer cancel()
	if err := s.db.PingContext(ctx); err != nil {
		problem(c, http.StatusServiceUnavailable, "PostgreSQL 不可用", err)
		return
	}
	if err := s.redis.Ping(ctx).Err(); err != nil {
		problem(c, http.StatusServiceUnavailable, "Redis 不可用", err)
		return
	}
	c.JSON(http.StatusOK, gin.H{"status": "ready"})
}

func (s *Server) requestID() gin.HandlerFunc {
	return func(c *gin.Context) {
		id := c.GetHeader("X-Request-ID")
		if id == "" {
			id = uuid.NewString()
		}
		c.Set("request_id", id)
		c.Header("X-Request-ID", id)
		c.Next()
	}
}

func (s *Server) securityHeaders() gin.HandlerFunc {
	return func(c *gin.Context) {
		c.Header("X-Content-Type-Options", "nosniff")
		c.Header("X-Frame-Options", "DENY")
		c.Header("Referrer-Policy", "no-referrer")
		c.Header("Permissions-Policy", "camera=(), microphone=(), geolocation=()")
		c.Header("Content-Security-Policy", "default-src 'self'; connect-src 'self'; img-src 'self' data:; style-src 'self' 'unsafe-inline'; script-src 'self'")
		c.Next()
	}
}

func (s *Server) accessLog() gin.HandlerFunc {
	return func(c *gin.Context) {
		started := time.Now()
		c.Next()
		fields := []zap.Field{zap.String("method", c.Request.Method), zap.String("path", c.Request.URL.Path), zap.Int("status", c.Writer.Status()), zap.Duration("duration", time.Since(started)), zap.String("request_id", c.GetString("request_id"))}
		if len(c.Errors) > 0 {
			fields = append(fields, zap.String("errors", c.Errors.String()))
		}
		s.logger.Info("http request", fields...)
	}
}

func (s *Server) serveSPA(c *gin.Context) {
	if c.Request.Method != http.MethodGet && c.Request.Method != http.MethodHead {
		c.Status(http.StatusNotFound)
		return
	}
	path := strings.TrimPrefix(c.Request.URL.Path, "/")
	if path != "" {
		if _, err := fs.Stat(s.assets, path); err == nil {
			if strings.HasPrefix(path, "assets/") {
				c.Header("Cache-Control", "public, max-age=31536000, immutable")
			}
			http.FileServer(http.FS(s.assets)).ServeHTTP(c.Writer, c.Request)
			return
		}
	}
	c.Header("Cache-Control", "no-cache")
	data, err := fs.ReadFile(s.assets, "index.html")
	if err != nil {
		problem(c, http.StatusNotFound, "页面不存在", err)
		return
	}
	c.Data(http.StatusOK, "text/html; charset=utf-8", data)
}